Split boundary event tests
