+++ _integration_tests/test/boundary_event_error_test.js
