+++ _integration_tests/test/boundary_event_timer_test.js
